fix: remove docker layer caching
Docker caching only works if file contents change as part of the image
and aren't affected by changes created by scripts.

Since this image is based on tool installs we should always rebuild the
image without a cache. The image isn't built too often now so it won't
really be an impact